requirements.txt
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
wordcloud==1.8.2.2
zipp==3.8.1
